fix: 변경 SearchDetailFragment -> Activity
- 화면 비율과 인스턴스 저장을 위해 activity로 변경
- activity 변경으로 인한 intent 형식 변경
