+++ app/src/main/AndroidManifest.xml
@@ -17,6 +17,8 @@
         android:theme="@style/Theme.FindHomes"
         tools:targetApi="31">
         <activity
+            android:name=".presentation.ui.search.SearchDetailActivity"/>
+        <activity
             android:name=".presentation.ui.login.LogInActivity"
             android:exported="false" />
         <activity
