+++ app/src/main/res/layout/activity_search_detail.xml
@@ -2,7 +2,8 @@
 <layout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".presentation.ui.search.SearchDetailFragment">
+    android:background="@color/white"
+    tools:context=".presentation.ui.search.SearchDetailActivity">
 
     <data>
         <import type="android.view.View"/>
